FIX: Catching Exception instead of the a specific one in DatabaseUpdater
With this commit, all possible Exceptions are caught. This prevents the application from crashing, especially if the enteres domain and port are wrong and the host refuses the connection

Some unnecessary changes from last commit are also reverted, such as the change in IHttpClient. Furthermore a test is written to cover the catch clause.

Signed-off-by: mkasimd <49522425+mkasimd@users.noreply.github.com>